==> scripts/1_required_packages.R <==
#________________________________________________#
##### Install libraries ##########################
#________________________________________________#

remotes::install_version('unmarked', '1.3.2') # to extract confint, need v.1.3.2
install.packages('tidyverse')
install.packages('grid')
install.packages('gridExtra')
install.packages('magick')
